Seems ok for me

+++ src/hash.cpp
@@ -65,7 +65,7 @@ auto split_piece_hashes(const std::string& piece_hashes_str) -> HashResult {
 }
 
 bool verify_piece(const std::vector<uint8_t>& piece, hash_t hash) {
-  assert(piece.size() < std::numeric_limits<int>::max());
+  assert(piece.size() < std::numeric_limits<size_t>::max());
   hash_t buffer;
   sha1::calc(piece.data(), static_cast<int>(piece.size()), buffer.begin());
   return buffer == hash;
